Remove "Network" path as required dir to run automactc.py on mounted image
"Network" path does not exist as an actual path on root for an OSX. Also, not everyone will have a network share or path mounted. Having this check essentially makes it impossible to run the automactc.py on a mounted image without removing that check. The current required_dirs seem sufficient to identify a typical OSX filesystem, however other directories could be added (They will be hidden) such as "private", "usr", or "sbin".